fix: preserve full stats week buckets
