+++ web/src/pages/InsightsPage.tsx
@@ -17,10 +17,6 @@ import { usePageHeader } from "@/contexts/PageHeaderContext";
 import PageLayout from "@/layouts/PageLayout";
 import { formatDate, formatDuration } from "@/utils/datetime";
 import { Icon } from "@/components/icons";
-import type {
-  CalendarAdapter,
-  ExtendedPlanningViewType,
-} from "@/utils/calendar";
 import type { Area } from "@/services/api/areas";
 import ActionButton from "@/components/ActionButton";
 import { RadioGroup, SegmentedControl, TextInput } from "@/components/forms";
@@ -40,6 +36,10 @@ import {
   buildPeriodCoverage,
   type PeriodCoverage,
 } from "@/features/insights/periodCoverage";
+import {
+  buildBucketBoundaries,
+  parseLocalDate,
+} from "@/features/insights/periodBuckets";
 
 type AreaBucket = {
   areaId: UUID;
@@ -62,26 +62,6 @@ const INSIGHT_VIEW_CONFIG: InsightViewConfig = {
 
 const MS_PER_DAY = 24 * 60 * 60 * 1000;
 
-const parseLocalDate = (isoDate: string): Date => {
-  const [yearStr, monthStr, dayStr] = isoDate.split("-");
-  const year = Number(yearStr);
-  const month = Number(monthStr) - 1;
-  const day = Number(dayStr);
-  return new Date(year, month, day, 0, 0, 0, 0);
-};
-
-const toIsoDate = (date: Date): string => {
-  const normalized = new Date(date);
-  normalized.setHours(0, 0, 0, 0);
-  return normalized.toLocaleDateString("en-CA");
-};
-
-const normalizeDate = (date: Date): Date => {
-  const normalized = new Date(date);
-  normalized.setHours(0, 0, 0, 0);
-  return normalized;
-};
-
 const calculateInclusiveDays = (start: string, end?: string | null): number => {
   if (!start) return 0;
   const startDate = parseLocalDate(start);
@@ -92,97 +72,6 @@ const calculateInclusiveDays = (start: string, end?: string | null): number => {
   const upper = Math.max(startTime, endTime);
   const diff = upper - lower;
   return Math.max(1, Math.floor(diff / MS_PER_DAY) + 1);
-};
-
-const buildBucketBoundaries = (
-  granularity: AggregationGranularity,
-  startDate: string,
-  endDate: string,
-  calendarAdapter: CalendarAdapter,
-): Array<{ start: string; end: string }> => {
-  if (!startDate || !endDate) return [];
-  const buckets: Array<{ start: string; end: string }> = [];
-  const rangeStart = parseLocalDate(startDate);
-  const rangeEnd = parseLocalDate(endDate);
-  const toIso = (d: Date) => toIsoDate(normalizeDate(d));
-
-  const appendBucket = (start: Date, end: Date) => {
-    const normalizedStart = normalizeDate(start);
-    const normalizedEnd = normalizeDate(end);
-    if (normalizedStart > rangeEnd) return;
-    const clampedStart =
-      normalizedStart < rangeStart ? rangeStart : normalizedStart;
-    const clampedEnd = normalizedEnd > rangeEnd ? rangeEnd : normalizedEnd;
-    if (clampedStart > clampedEnd) return;
-    buckets.push({ start: toIso(clampedStart), end: toIso(clampedEnd) });
-  };
-
-  if (granularity === "day") {
-    for (
-      let cursor = new Date(rangeStart);
-      cursor <= rangeEnd;
-      cursor.setDate(cursor.getDate() + 1)
-    ) {
-      const iso = toIsoDate(cursor);
-      buckets.push({ start: iso, end: iso });
-    }
-    return buckets;
-  }
-
-  const viewTypeMap: Record<
-    Exclude<AggregationGranularity, "day">,
-    ExtendedPlanningViewType
-  > = {
-    week: "week",
-    month: "month",
-    year: "year",
-  };
-
-  const viewType =
-    viewTypeMap[granularity as Exclude<AggregationGranularity, "day">];
-  const parseRange = (range: { start: string; end: string }) => ({
-    start: parseLocalDate(range.start),
-    end: parseLocalDate(range.end),
-  });
-
-  let currentRange = calendarAdapter.getPeriodRange(viewType, rangeStart);
-  let { start, end } = parseRange(currentRange);
-  let safety = 0;
-
-  while (end < rangeStart && safety < 500) {
-    const nextRange = calendarAdapter.shiftPeriodRange(
-      viewType,
-      currentRange.start,
-      currentRange.end,
-      1,
-    );
-    currentRange = nextRange;
-    ({ start, end } = parseRange(currentRange));
-    safety += 1;
-  }
-
-  while (start <= rangeEnd && safety < 1500) {
-    appendBucket(start, end);
-    if (end >= rangeEnd) break;
-
-    const nextRange = calendarAdapter.shiftPeriodRange(
-      viewType,
-      currentRange.start,
-      currentRange.end,
-      1,
-    );
-    const nextParsed = parseRange(nextRange);
-    if (nextParsed.start.getTime() === start.getTime()) {
-      // Prevent infinite loops if adapter returns the same range
-      break;
-    }
-    currentRange = nextRange;
-    start = nextParsed.start;
-    end = nextParsed.end;
-    safety += 1;
-  }
-
-  return buckets;
 };
 
 const calculateBucketCapacityMinutes = (start: string, end: string): number => {
